+++ src/_types_.js
@@ -1,0 +1,25 @@
+/**
+ * Mask.
+ *
+ * @typedef  Mask
+ * @property {number} x      - x
+ * @property {number} y      - y
+ * @property {number} width  - width
+ * @property {number} height - height
+ */
+/**
+ * MaskRadius.
+ *
+ * @typedef  MaskRadius
+ * @property {number} topLeft     - topLeft
+ * @property {number} topRight    - topRight
+ * @property {number} bottomRight - bottomRight
+ * @property {number} bottomLeft  - bottomLeft
+ */
+/**
+ * Position.
+ *
+ * @typedef  Position
+ * @property {number} x - x
+ * @property {number} y - y
+ */
