+++ tests/uploader.fn.load_image.test.js
@@ -23,6 +23,7 @@ window.namespace.draw = function() {};
 describe("uploader", function(){
     beforeEach(function() {
         require("./required.js");
+        require("../src/helpers");
         require("../src/uploader");
     });
