+++ tests/uploader.fn.move.test.js
@@ -29,6 +29,7 @@ window.customZoomUpdate = function(uploaderInstance, from) {
 describe("uploader", function(){
     beforeEach(function() {
         require("./required.js");
+        require("../src/helpers");
         require("../src/uploader");
     });
