+++ tests/uploader.fn.reset.test.js
@@ -10,6 +10,7 @@ window.cancel = function() {};
 describe("uploader", function(){
     beforeEach(function() {
         require("./required.js");
+        require("../src/helpers");
         require("../src/uploader");
     });
